Fixed some failing specs from the previous commit. Added a user model and changed all occurances of ip strings to use the usder model instead.

+++ models/nomination.rb
@@ -6,11 +6,11 @@ class Nomination
   property :id, Serial
   property :score, Integer, :default => 0
   property :status, String, :length => 20
-  property :nominated_by, String
   property :played_at, DateTime
   property :created_at, DateTime
     
   belongs_to :album
+  belongs_to :user
   has n, :songs, :through => Resource
   has n, :votes, :type => "vote", :value.gt => 0
   has n, :negative_votes, :class_name => "Vote", :type => "vote", :value.lt => 0
@@ -20,18 +20,22 @@ class Nomination
   def artist; album.artist end
   def name; album.name end
   
-  def owned_by?(ip); nominated_by == ip end
+  def owned_by?(ip); user && user.ip == ip end
   
   # Song management methods
   # ----------------------------------------------------------------------
-  def add(song_id)
-    song = album.songs.get(song_id)
-    return unless song && !songs.include?(song)
+  def add(song_id, ip)
+    return unless owned_by?(ip)
     
-    self.songs << song
-    save
+    song = album.songs.get(song_id)
+    if song && !songs.include?(song)
+      self.songs << song
+      save
+    end
   end
-  def delete(song_id)
+  def delete(song_id, ip)
+    return unless owned_by?(ip)
+    
     song = NominationSong.first(:nomination_id => id, :song_id => song_id)
     song.destroy if song
   end
@@ -42,11 +46,11 @@ class Nomination
     return unless can_be_voted_for_by?(ip)
     
     self.score = score + value
-    self.send(value > 0 ? :votes : :negative_votes).create :value => value, :ip => ip, :type => "vote"
+    self.send(value > 0 ? :votes : :negative_votes).create :user => User.get_or_create_by(ip), :value => value, :type => "vote"
     self.status = "deleted" if score <= DEFAULT_ELIMINATION_SCORE
     save
   end
-  def can_be_voted_for_by?(ip); !(votes + negative_votes).map { |v| v.ip }.include?(ip) end
+  def can_be_voted_for_by?(ip); !(votes + negative_votes).map { |v| v.user }.include?(User.get_or_create_by(ip)) end
   
   def remove(ip); self.update_attributes(:status => "deleted") if owned_by?(ip) end
 
@@ -56,18 +60,21 @@ class Nomination
   def force(ip)
     return unless can_be_forced_by?(ip)
     
-    self.down_votes.create :value => 1, :ip => ip, :type => "force"
+    self.down_votes.create :user => User.get_or_create_by(ip), :value => 1, :type => "force"
     MpdProxy.execute(:clear) if down_votes_necessary <= 0
   end
-  def can_be_forced_by?(ip); !down_votes.map { |v| v.ip }.include?(ip) && negative_votes.map { |v| v.ip }.include?(ip) end
+  def can_be_forced_by?(ip)
+    user = User.get_or_create_by(ip)
+    !down_votes.map { |v| v.user }.include?(user) && negative_votes.map { |v| v.user }.include?(user)
+  end
 
   # Rate methods
   # ----------------------------------------------------------------------
   def rate(value, ip)
-    return if ratings.map { |v| v.ip }.include?(ip)
-    self.ratings.create :value => [[value, 1].max, 5].min, :ip => ip, :type => "rating"
+    return unless can_be_rated_by?(ip)
+    self.ratings.create :user => User.get_or_create_by(ip), :value => [[value, 1].max, 5].min, :type => "rating"
   end
-  def can_be_rated_by?(ip); !ratings.map { |v| v.ip }.include?(ip) end
+  def can_be_rated_by?(ip); !ratings.map { |v| v.user }.include?(User.get_or_create_by(ip)) end
   
   # Class methods
   # ----------------------------------------------------------------------
